+++ src/main/java/net/frozenblock/wilderwild/config/WWMixinsConfig.java
@@ -73,8 +73,6 @@ public final class WWMixinsConfig {
 
 	public boolean block_dripleaf = true;
 
-	public boolean block_dripstone = true;
-
 	public boolean block_echo_glass = true;
 
 	public boolean block_fire = true;
@@ -112,8 +110,6 @@ public final class WWMixinsConfig {
 	public boolean entity_experience = true;
 
 	public boolean entity_firefly = true;
-
-	public boolean entity_jellyfish = true;
 
 	public boolean entity_lightning = true;