+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBlockTagProvider.java
@@ -148,8 +148,6 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 	private void generateLib() {
 		this.getOrCreateTagBuilder(FrozenBlockTags.DRIPSTONE_CAN_DRIP_ON)
 			.add(Blocks.DIRT)
-			.add(Blocks.SAND)
-			.add(Blocks.RED_SAND)
 			.add(WWBlocks.SCORCHED_SAND)
 			.add(WWBlocks.SCORCHED_RED_SAND);
 	}
@@ -526,7 +524,8 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.WAXED_COPPER_GRATE)
 			.add(Blocks.WAXED_EXPOSED_COPPER_GRATE)
 			.add(Blocks.WAXED_WEATHERED_COPPER_GRATE)
-			.add(Blocks.WAXED_OXIDIZED_COPPER_GRATE);
+			.add(Blocks.WAXED_OXIDIZED_COPPER_GRATE)
+			.addOptionalTag(WWBlockTags.MESOGLEA);
 
 		this.getOrCreateTagBuilder(WWBlockTags.GEYSER_CANNOT_PASS_THROUGH)
 			.addOptionalTag(ConventionalBlockTags.GLASS_BLOCKS);