+++ src/main/java/net/frozenblock/wilderwild/entity/Butterfly.java
@@ -25,7 +25,7 @@ import java.util.Optional;
 import net.frozenblock.lib.wind.api.WindManager;
 import net.frozenblock.wilderwild.config.WWEntityConfig;
 import net.frozenblock.wilderwild.entity.ai.butterfly.ButterflyAi;
-import net.frozenblock.wilderwild.entity.impl.Bottleable;
+import net.frozenblock.wilderwild.entity.impl.WWBottleable;
 import net.frozenblock.wilderwild.entity.variant.butterfly.ButterflyVariant;
 import net.frozenblock.wilderwild.entity.variant.butterfly.ButterflyVariants;
 import net.frozenblock.wilderwild.registry.WWDataComponents;
@@ -86,7 +86,7 @@ import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable {
+public class Butterfly extends PathfinderMob implements FlyingAnimal, WWBottleable {
 	public static final int TICKS_PER_FLAP = 3;
 	public static final int SPAWN_CHANCE = 50;
 	private static final double SPAWN_RADIUS_CHECK_DISTANCE = 20D;
@@ -204,7 +204,7 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 	@Override
 	@NotNull
 	protected InteractionResult mobInteract(@NotNull Player player, @NotNull InteractionHand hand) {
-		return Bottleable.bottleMobPickup(player, hand, this).orElse(super.mobInteract(player, hand));
+		return WWBottleable.bottleMobPickup(player, hand, this).orElse(super.mobInteract(player, hand));
 	}
 
 	@Override
@@ -225,28 +225,28 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 	}
 
 	@Override
-	public boolean fromBottle() {
+	public boolean wilderWild$fromBottle() {
 		return this.entityData.get(FROM_BOTTLE);
 	}
 
 	@Override
-	public void setFromBottle(boolean value) {
+	public void wilderWild$setFromBottle(boolean value) {
 		this.entityData.set(FROM_BOTTLE, value);
 	}
 
 	@Override
-	public void saveToBottleTag(ItemStack itemStack) {
-		Bottleable.saveDefaultDataToBottleTag(this, itemStack);
+	public void wilderWild$saveToBottleTag(ItemStack itemStack) {
+		WWBottleable.saveDefaultDataToBottleTag(this, itemStack);
 		itemStack.copyFrom(WWDataComponents.BUTTERFLY_VARIANT, this);
 	}
 
 	@Override
-	public void loadFromBottleEntityDataTag(CompoundTag compoundTag) {
-		Bottleable.loadDefaultDataFromBottleTag(this, compoundTag);
+	public void wilderWild$loadFromBottleEntityDataTag(CompoundTag compoundTag) {
+		WWBottleable.loadDefaultDataFromBottleTag(this, compoundTag);
 	}
 
 	@Override
-	public void loadFromBottleItemStack(@NotNull ItemStack itemStack) {
+	public void wilderWild$loadFromBottleItemStack(@NotNull ItemStack itemStack) {
 		Holder<ButterflyVariant> variant = itemStack.getOrDefault(WWDataComponents.BUTTERFLY_VARIANT, this.registryAccess().getOrThrow(ButterflyVariants.DEFAULT));
 		this.setVariant(variant.value());
 	}
@@ -277,21 +277,21 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 	}
 
 	@Override
-	public void onCapture() {
+	public void wilderWild$onCapture() {
 	}
 
 	@Override
-	public void onBottleRelease() {
+	public void wilderWild$onBottleRelease() {
 		ButterflyAi.rememberHome(this, this.blockPosition());
 	}
 
 	@Override
-	public ItemStack getBottleItemStack() {
+	public ItemStack wilderWild$getBottleItemStack() {
 		return new ItemStack(WWItems.BUTTERFLY_BOTTLE);
 	}
 
 	@Override
-	public SoundEvent getBottleCatchSound() {
+	public SoundEvent wilderWild$getBottleCatchSound() {
 		return WWSounds.ITEM_BOTTLE_CATCH_BUTTERFLY;
 	}
 
@@ -325,7 +325,7 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 
 	@Override
 	public boolean requiresCustomPersistence() {
-		return super.requiresCustomPersistence() || this.fromBottle();
+		return super.requiresCustomPersistence() || this.wilderWild$fromBottle();
 	}
 
 	@Override
@@ -428,14 +428,14 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 
 	@Override
 	public boolean removeWhenFarAway(double distanceToClosestPlayer) {
-		return !this.fromBottle() && !this.hasCustomName();
+		return !this.wilderWild$fromBottle() && !this.hasCustomName();
 	}
 
 	@Override
 	public void addAdditionalSaveData(@NotNull CompoundTag compoundTag) {
 		super.addAdditionalSaveData(compoundTag);
 		VariantUtils.writeVariant(compoundTag, this.getVariantAsHolder());
-		compoundTag.putBoolean("fromBottle", this.fromBottle());
+		compoundTag.putBoolean("fromBottle", this.wilderWild$fromBottle());
 	}
 
 	@Override
@@ -443,7 +443,7 @@ public class Butterfly extends PathfinderMob implements FlyingAnimal, Bottleable
 		super.readAdditionalSaveData(compoundTag);
 		VariantUtils.readVariant(compoundTag, this.registryAccess(), WilderWildRegistries.BUTTERFLY_VARIANT)
 			.ifPresent(butterflyVariantHolder -> this.setVariant(butterflyVariantHolder.value()));
-		if (compoundTag.contains("fromBottle")) this.setFromBottle(compoundTag.getBoolean("fromBottle"));
+		if (compoundTag.contains("fromBottle")) this.wilderWild$setFromBottle(compoundTag.getBoolean("fromBottle"));
 	}
 
 	@Override