+++ src/main/java/net/frozenblock/wilderwild/entity/Firefly.java
@@ -26,7 +26,7 @@ import java.util.Optional;
 import net.frozenblock.lib.wind.api.WindManager;
 import net.frozenblock.wilderwild.config.WWEntityConfig;
 import net.frozenblock.wilderwild.entity.ai.firefly.FireflyAi;
-import net.frozenblock.wilderwild.entity.impl.Bottleable;
+import net.frozenblock.wilderwild.entity.impl.WWBottleable;
 import net.frozenblock.wilderwild.entity.variant.firefly.FireflyColor;
 import net.frozenblock.wilderwild.entity.variant.firefly.FireflyColors;
 import net.frozenblock.wilderwild.registry.WWDataComponents;
@@ -86,7 +86,7 @@ import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
+public class Firefly extends PathfinderMob implements FlyingAnimal, WWBottleable {
 	public static final int RANDOM_FLICKER_AGE_MAX = 19;
 	private static final EntityDataAccessor<Boolean> FROM_BOTTLE = SynchedEntityData.defineId(Firefly.class, EntityDataSerializers.BOOLEAN);
 	private static final EntityDataAccessor<Integer> AGE = SynchedEntityData.defineId(Firefly.class, EntityDataSerializers.INT);
@@ -202,7 +202,7 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 	@Override
 	@NotNull
 	protected InteractionResult mobInteract(@NotNull Player player, @NotNull InteractionHand hand) {
-		return Bottleable.bottleMobPickup(player, hand, this).orElse(super.mobInteract(player, hand));
+		return WWBottleable.bottleMobPickup(player, hand, this).orElse(super.mobInteract(player, hand));
 	}
 
 	@Override
@@ -228,28 +228,28 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 	}
 
 	@Override
-	public boolean fromBottle() {
+	public boolean wilderWild$fromBottle() {
 		return this.entityData.get(FROM_BOTTLE);
 	}
 
 	@Override
-	public void setFromBottle(boolean value) {
+	public void wilderWild$setFromBottle(boolean value) {
 		this.entityData.set(FROM_BOTTLE, value);
 	}
 
 	@Override
-	public void saveToBottleTag(ItemStack itemStack) {
-		Bottleable.saveDefaultDataToBottleTag(this, itemStack);
+	public void wilderWild$saveToBottleTag(ItemStack itemStack) {
+		WWBottleable.saveDefaultDataToBottleTag(this, itemStack);
 		itemStack.copyFrom(WWDataComponents.FIREFLY_COLOR, this);
 	}
 
 	@Override
-	public void loadFromBottleEntityDataTag(@NotNull CompoundTag compoundTag) {
-		Bottleable.loadDefaultDataFromBottleTag(this, compoundTag);
+	public void wilderWild$loadFromBottleEntityDataTag(@NotNull CompoundTag compoundTag) {
+		WWBottleable.loadDefaultDataFromBottleTag(this, compoundTag);
 	}
 
 	@Override
-	public void loadFromBottleItemStack(@NotNull ItemStack itemStack) {
+	public void wilderWild$loadFromBottleItemStack(@NotNull ItemStack itemStack) {
 		Holder<FireflyColor> color = itemStack.getOrDefault(WWDataComponents.FIREFLY_COLOR, this.registryAccess().getOrThrow(FireflyColors.DEFAULT));
 		this.setColor(color.value());
 	}
@@ -280,24 +280,24 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 	}
 
 	@Override
-	public void onCapture() {
+	public void wilderWild$onCapture() {
 		if (this.isSwarmLeader()) {
 			FireflyAi.transferLeadershipToRandomFirefly(this);
 		}
 	}
 
 	@Override
-	public void onBottleRelease() {
+	public void wilderWild$onBottleRelease() {
 		FireflyAi.rememberHome(this, this.blockPosition());
 	}
 
 	@Override
-	public ItemStack getBottleItemStack() {
+	public ItemStack wilderWild$getBottleItemStack() {
 		return new ItemStack(WWItems.FIREFLY_BOTTLE);
 	}
 
 	@Override
-	public SoundEvent getBottleCatchSound() {
+	public SoundEvent wilderWild$getBottleCatchSound() {
 		return WWSounds.ITEM_BOTTLE_CATCH_FIREFLY;
 	}
 
@@ -355,7 +355,7 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 
 	@Override
 	public boolean requiresCustomPersistence() {
-		return super.requiresCustomPersistence() || this.fromBottle();
+		return super.requiresCustomPersistence() || this.wilderWild$fromBottle();
 	}
 
 	@Override
@@ -504,7 +504,7 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 
 	@Override
 	public boolean removeWhenFarAway(double distanceToClosestPlayer) {
-		return !this.fromBottle() && !this.hasCustomName();
+		return !this.wilderWild$fromBottle() && !this.hasCustomName();
 	}
 
 	@Override
@@ -523,7 +523,7 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 			.unwrapKey()
 			.ifPresent(resourceKey -> compoundTag.putString("color", resourceKey.location().toString()));
 
-		compoundTag.putBoolean("fromBottle", this.fromBottle());
+		compoundTag.putBoolean("fromBottle", this.wilderWild$fromBottle());
 		compoundTag.putInt("flickerAge", this.getFlickerAge());
 		compoundTag.putFloat("scale", this.getAnimScale());
 		compoundTag.putFloat("prevScale", this.getPrevAnimScale());
@@ -538,7 +538,7 @@ public class Firefly extends PathfinderMob implements FlyingAnimal, Bottleable {
 			.flatMap(resourceKey -> this.registryAccess().lookupOrThrow(WilderWildRegistries.FIREFLY_COLOR).get(resourceKey))
 			.ifPresent(reference -> this.setColor(reference.value()));
 
-		if (compoundTag.contains("fromBottle")) this.setFromBottle(compoundTag.getBoolean("fromBottle"));
+		if (compoundTag.contains("fromBottle")) this.wilderWild$setFromBottle(compoundTag.getBoolean("fromBottle"));
 		if (compoundTag.contains("flickerAge")) this.setFlickerAge(compoundTag.getInt("flickerAge"));
 		if (compoundTag.contains("scale")) this.setAnimScale(compoundTag.getFloat("scale"));
 		if (compoundTag.contains("prevScale")) this.setPrevAnimScale(compoundTag.getFloat("prevScale"));