+++ src/main/java/net/frozenblock/wilderwild/entity/FlowerCow.java
@@ -53,7 +53,7 @@ import net.minecraft.world.entity.EntitySpawnReason;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Shearable;
 import net.minecraft.world.entity.SpawnGroupData;
-import net.minecraft.world.entity.animal.Cow;
+import net.minecraft.world.entity.animal.AbstractCow;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.variant.SpawnContext;
@@ -74,7 +74,7 @@ import net.minecraft.world.level.gameevent.GameEvent;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class FlowerCow extends Cow implements Shearable {
+public class FlowerCow extends AbstractCow implements Shearable {
 	public static final int MAX_FLOWERS = 4;
 	private static final byte GROW_FLOWER_EVENT_ID = 61;
 	private static final EntityDataAccessor<String> VARIANT = SynchedEntityData.defineId(FlowerCow.class, EntityDataSerializers.STRING);