+++ CHANGELOG.md
@@ -91,7 +91,12 @@ hi
   - This change was made to make Jellyfish more peaceful and realistic.
 - Mesoglea is now always "waterlogged," maaking it much easier for players to utilize it properly.
   - Mesoglea will no longer leave a water block behind when broken.
-  - Fluid-like rendering for Mesoglea is now enabled by default in the config, and has been optimized.
+  - Fluid-like rendering for Mesoglea has been optimized.
+  - Water no longer renders inside Mesoglea Blocks.
+  - Mesoglea now overrides the water fog color with its own.
+  - Geysers can now erupt through Mesoglea.
+  - Mesoglea now has custom Bubble, Bubble Pop, and Splash particles.
+    - Mesoglea Bubbles float and can persist in the air for a short period of time.
 - Rattling sounds are now used as a footstep sound instead of idle sound for Crabs.
 - Added the Phlox flower.
   - Generates in Birch Forest, Old Growth Birch Forest, and Meadow biomes.
@@ -104,6 +109,7 @@ hi
 - Magmatic Caves are now larger on average.
 - Lowered the average amount of Bushes that generate.
 - Increased the average amount of Tumbleweed that generate.
+- Sand and Red Sand can no longer be scorched by Lava falling from Pointed Dripstone.
 
 # Music
 - Added three new tracks by Willow/pictochats_.